+++ src/briefy/leica/models/professional/link.py
@@ -16,9 +16,11 @@ class Link(mixins.LeicaMixin, Base):
 
     _workflow = workflows.LinkWorkflow
 
-    __exclude_attributes__ = ['professional', ]
+    __default_exclude_attributes__ = ['request', 'versions', 'local_roles', 'path', 'can_view']
 
-    __summary_attributes__ = ['id', 'url']
+    __exclude_attributes__ = ['professional']
+
+    __summary_attributes__ = ['id', 'url', 'type', 'state']
 
     __summary_attributes_relations__ = []